fix: readd clicking outside of root to zoom out
